requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4